refactor: move postgres codec initialization to a separate file
